adjusted ambient lighting
in my past "Adventure on Mystery Island" project, I found that a sky contribution of 1 on the ambient light made it completely dark on integrated graphics. changing sky contribution to 0.5 or even 0.0, with a gray color as the ambient light makes it visible regardless. (essential because of how little extra lights i have)